Fix calculate_all_attentions
- Fix for attention plot for RNN
- Remove AbsAttention
- Move calculate_all_attentions.py: espnet2/train -> espnet2/main_funcs
